feat: cambia el tipo de coordenada de double a int

+++ trabajo-especial/microservice-monopatin/src/main/java/com/grupo08/unicen/microservicemonopatin/entity/Monopatin.java
@@ -19,8 +19,8 @@ public class Monopatin {
 
     @Enumerated(EnumType.STRING)
     private State state;
-    private Double x;
-    private Double y;
+    private int x;
+    private int y;
     private BigDecimal kmTraveled;
     private BigDecimal useTime;
 }
